--- antarest/study/storage/rawstudy/io/reader/__init__.py
@@ -1,3 +1,0 @@
-from antarest.study.storage.rawstudy.io.reader.ini_reader import IniReader, MultipleSameKeysIniReader
-
-__all__ = ["IniReader", "MultipleSameKeysIniReader"]
