+++ antarest/study/storage/rawstudy/model/filesystem/ini_file_node.py
@@ -13,8 +13,7 @@ from typing import Any, Callable, Dict, List, Optional, Union, cast
 from filelock import FileLock
 
 from antarest.core.model import JSON, SUB_JSON
-from antarest.study.storage.rawstudy.io.reader import IniReader
-from antarest.study.storage.rawstudy.io.reader.ini_reader import IReader
+from antarest.study.storage.rawstudy.io.reader.ini_reader import IniReader, IReader
 from antarest.study.storage.rawstudy.io.writer.ini_writer import IniWriter
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.context import ContextServer
