+++ antarest/study/storage/rawstudy/model/filesystem/root/settings/generaldata.py
@@ -1,7 +1,7 @@
 from copy import deepcopy
 from typing import Any, Dict
 
-from antarest.study.storage.rawstudy.io.reader import MultipleSameKeysIniReader
+from antarest.study.storage.rawstudy.io.reader.ini_reader import MultipleSameKeysIniReader
 from antarest.study.storage.rawstudy.io.writer.ini_writer import IniWriter
 from antarest.study.storage.rawstudy.model.filesystem.config.model import FileStudyTreeConfig
 from antarest.study.storage.rawstudy.model.filesystem.context import ContextServer
