+++ antarest/study/storage/study_upgrader/upgrader_830.py
@@ -1,7 +1,7 @@
 import glob
 from pathlib import Path
 
-from antarest.study.storage.rawstudy.io.reader import MultipleSameKeysIniReader
+from antarest.study.storage.rawstudy.io.reader.ini_reader import MultipleSameKeysIniReader
 from antarest.study.storage.rawstudy.io.writer.ini_writer import IniWriter
 from antarest.study.storage.rawstudy.model.filesystem.root.settings.generaldata import DUPLICATE_KEYS
