+++ tests/storage/repository/antares_io/reader/test_ini_reader.py
@@ -3,8 +3,7 @@ from pathlib import Path
 
 import pytest
 
-from antarest.study.storage.rawstudy.io.reader import IniReader
-from antarest.study.storage.rawstudy.io.reader.ini_reader import MultipleSameKeysIniReader
+from antarest.study.storage.rawstudy.io.reader.ini_reader import IniReader, MultipleSameKeysIniReader
 
 
 class TestIniReader:
