+++ tests/variantstudy/model/command/test_create_link.py
@@ -5,7 +5,7 @@ import pytest
 from pydantic import ValidationError
 
 from antarest.study.common.default_values import FilteringOptions, LinkProperties
-from antarest.study.storage.rawstudy.io.reader import IniReader
+from antarest.study.storage.rawstudy.io.reader.ini_reader import IniReader
 from antarest.study.storage.rawstudy.model.filesystem.config.model import transform_name_to_id
 from antarest.study.storage.rawstudy.model.filesystem.factory import FileStudy
 from antarest.study.storage.variantstudy.business.command_reverter import CommandReverter
